cni: Adjust Cilium ConfigMap to version 1.6.6
Cilium 1.6.6 brings new configuration options which need to be used in
order to make it working on CaaSP. It also brings the possibility of
using CRD instead of etcd for storing internal data. etcd should be used
only when using clustermesh. Skuba used the internal etcd of the single
Kubernetes cluster for Cilium, which already excludes the clustermesh
case. Since we don't support clustermesh yet, using CRD will be the
default option for Cilium 1.6.

Signed-off-by: Michal Rostecki <mrostecki@suse.de>