addons: Perform migration of Cilium data from etcd to CRD
Cilium from version 1.6 can use CRD for storing internal data and using
etcd is not needed anymore when not running clustermesh. Skuba was using
the internal etcd cluster of the single Kubernetes cluster for Cilium,
which already exludes the clustermesh case. Since we don't support
clustermesh yet, for now skuba is going to use CRD in every case and
migrate data from old Cilium deployments.

The process of migration is not mandatory, but without it, Cilium is
going to regenerate data from scratch which might result in service
downtimes. If the automated migration is not successful, the upgrade
should still proceed with the risk of temporary service downtimes.

Signed-off-by: Michal Rostecki <mrostecki@suse.de>